+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - Added `maxBins` hyper-parameter to CART-based learners
     - Added stream Deduplicator extractor
     - Added the SiLU activation function
+    - Added Swish activation layer
     
 - 1.2.3
     - Fix Multiclass layer cross entropy gradient optimization