+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- 1.2.4
+    - Refactor neural network parameter updates
+
 - 1.2.3
     - Fix Multiclass layer cross entropy gradient optimization